requirements.txt
numpy
vosk
llama-cpp-python
pyaudio
piper-tts
scikit-learn
psutil
sounddevice

